chore: locate all tests in same folder
