+++ tests/Icons.svelte
