--- tests/import.test.ts
@@ -1,6 +1,0 @@
-import { test, expect } from "vitest";
-import * as icons from "../lib";
-
-test("imports", () => {
-  expect(Object.keys(icons).length).toEqual(1938);
-});
